+++ printers/plain.go
@@ -147,6 +147,7 @@ func (p *PlainPrinter) PrintStatistics(s *stats.Statistics) {
 
 	fmt.Fprintf(&msg, "total uptime:   %s\n", s.TotalUptimeDuration())
 	fmt.Fprintf(&msg, "total downtime: %s\n", s.TotalDowntimeDuration())
+	fmt.Fprintf(&msg, "outages: %d\n", s.TotalOutages)
 
 	if s.LongestUptime.Duration != 0 {
 		msg.WriteString("longest consecutive uptime:   ")